Struct keypom::Keypom

pub struct Keypom {

    pub owner_id: AccountId,
    pub root_account: AccountId,
    pub drop_id_for_pk: UnorderedMap<PublicKey, u128>,
    pub drop_for_id: LookupMap<u128, Drop>,
    pub drop_ids_for_owner: LookupMap<AccountId, UnorderedSet<u128>>,
    pub drop_fee: u128,
    pub key_fee: u128,
    pub fees_collected: u128,
    pub fees_per_user: LookupMap<AccountId, (u128, u128)>,
    pub user_balances: LookupMap<AccountId, Balance>,
    pub next_drop_id: u128,
    pub yocto_per_gas: u128,
    pub prohibited_fc_methods: LookupSet<String>,

}

Main contract struct that holds all the contract data

Fields
owner_id: AccountId

Owner of the Keypom contract that can call internal methods (e.g. set_fees)

root_account: AccountId

Which contract is the actual linkdrop deployed to (i.e testnet or near)

drop_id_for_pk: UnorderedMap<PublicKey, u128>

Map of each key to its respective drop ID. This is much more efficient than repeating the Drop data for every single key.

drop_for_id: LookupMap<u128, Drop>

Map a drop ID to its respective Drop data

drop_ids_for_owner: LookupMap<AccountId, UnorderedSet<u128>>

Keep track of the drop ids that each funder has created. This is used for view methods.

drop_fee: u128

Fee taken by the contract every time a new drop is created

key_fee: u128

Fee taken by the contract everytime a new key is added to a drop

fees_collected: u128

Total amount of fees available for withdrawal collected overtime.

fees_per_user: LookupMap<AccountId, (u128, u128)>

Overload the drop_fee and key_fee for specific users by providing custom fees Tuple is (drop_fee, key_fee)

user_balances: LookupMap<AccountId, Balance>

Keep track of the balances for each user. This is to prepay for drop creations

next_drop_id: u128

Keep track of a nonce used for the drop IDs

yocto_per_gas: u128

How many yoctoNEAR does 1 unit of Gas cost

prohibited_fc_methods: LookupSet<String>

Which methods are prohibited from being called with an access key through an FC Drop

Implementations

impl Keypom

pub fn add_prohibited_method(&mut self, method: String)

Add a prohibited method to the list of methods that can’t be called by a FC Drop

pub fn set_root_account(&mut self, root_account: AccountId)

Set the desired linkdrop contract to interact with

pub fn set_fees(&mut self, drop_fee: Option<U128>, key_fee: Option<U128>)

Set the desired contract fees

pub fn set_fees_per_user(
&mut self,
account_id: AccountId,
drop_fee: U128,
key_fee: U128
)

Set the desired linkdrop contract to interact with

pub fn set_gas_price(&mut self, yocto_per_gas: u128)

Set the desired linkdrop contract to interact with

pub fn withdraw_fees(&mut self, withdraw_to: AccountId) -> Promise

Withdraw the fees collected to the passed in Account Id

pub fn on_withdraw_fees(&mut self, fees_collected: u128) -> bool

Callback for withdrawing fees on the contract

impl Keypom

pub fn add_to_balance(&mut self)

pub fn withdraw_from_balance(&mut self)

pub fn get_user_balance(&self, account_id: AccountId) -> U128

impl Keypom

pub fn delete_keys(
&mut self,
drop_id: u128,
public_keys: Option<Vec<PublicKey>>,
limit: Option<u8>,
delete_on_empty: Option<bool>
)

pub fn refund_assets(&mut self, drop_id: u128, assets_to_refund: Option<u64>)

impl Keypom

pub fn create_drop(
&mut self,
public_keys: Vec<PublicKey>,
deposit_per_use: U128,
config: Option<DropConfig>,
metadata: Option<String>,
ft_data: Option<FTDataConfig>,
nft_data: Option<NFTDataConfig>,
fc_data: Option<FCData>
) -> u128

pub fn add_keys(&mut self, public_keys: Vec<PublicKey>, drop_id: u128) -> u128

impl Keypom

pub fn ft_on_transfer(
&mut self,
sender_id: AccountId,
amount: U128,
msg: U128
) -> PromiseOrValue<U128>

Allows users to attach fungible tokens to the Linkdrops. Must have storage recorded by this point. You can only attach one set of FTs or NFT at a time.

pub fn ft_resolve_batch(
&mut self,
amount: U128,
token_sender: AccountId,
token_contract: AccountId
) -> bool

Self callback checks if fungible tokens were successfully transferred to the new account. If yes, do nothing. If no, refund original sender

pub fn ft_resolve_refund(
&mut self,
drop_id: u128,
num_to_refund: u64,
ft_storage: U128,
funder: AccountId
) -> bool

Self callback checks if fungible tokens were successfully refunded. If yes, set keys registered to 0.

pub fn resolve_storage_check(
&mut self,
public_keys: Vec<PublicKey>,
drop_id: u128,
required_deposit: u128
) -> bool

self callback gets the storage balance bounds and inserts that into account data for each public key passed in

impl Keypom

pub fn nft_on_transfer(
&mut self,
token_id: String,
sender_id: AccountId,
msg: U128
) -> PromiseOrValue<bool>

pub fn nft_resolve_refund(
&mut self,
drop_id: U128,
token_ids: Vec<String>
) -> bool

self callback checks if NFT was successfully transferred to the new account. If yes, do nothing. If no, refund original sender

pub fn nft_resolve_transfer(
&mut self,
token_id: String,
token_sender: AccountId,
token_contract: AccountId
) -> bool

self callback checks if NFT was successfully transferred to the new account. If yes, do nothing. If no, refund original sender

impl Keypom

pub fn claim(&mut self, account_id: AccountId, expected_uses: Option<u64>)

Claim tokens for specific account that are attached to the public key this tx is signed with.

pub fn create_account_and_claim(
&mut self,
new_account_id: AccountId,
new_public_key: PublicKey,
expected_uses: Option<u64>
)

Create new account and and claim tokens to it.

pub fn on_claim_simple(
&mut self,
owner_id: AccountId,
balance: U128,
storage_used: Balance
) -> bool

self callback for simple linkdrops with no FTs, NFTs, or FCs.

pub fn on_claim_ft(
&mut self,
account_id: AccountId,
owner_id: AccountId,
balance: U128,
storage_used: Balance,
ft_data: FTData,
execute: bool
) -> bool

self callback for FT linkdrop

pub fn on_claim_nft(
&mut self,
account_id: AccountId,
owner_id: AccountId,
balance: U128,
storage_used: Balance,
sender_id: AccountId,
contract_id: AccountId,
token_id: String,
execute: bool
) -> bool

self callback for a linkdrop loaded with an NFT

pub fn on_claim_fc(
&mut self,
account_id: AccountId,
owner_id: AccountId,
balance: U128,
storage_used: Balance,
fc_data: FCData,
drop_id: u128,
cur_key_info: KeyInfo,
uses_per_key: u64,
execute: bool
) -> bool

self callback checks if account was created successfully or not. If yes, refunds excess storage, sends NFTs, FTs etc..

impl Keypom

pub fn get_key_balance(&self, key: PublicKey) -> U128

Returns the balance associated with given key. This is used by the NEAR wallet to display the amount of the linkdrop

pub fn get_key_total_supply(&self) -> U128

Query for the total supply of keys on the contract

pub fn get_keys(
&self,
from_index: Option<U128>,
limit: Option<u64>
) -> Vec<JsonKeyInfo>

Paginate through all active keys on the contract and return a vector of key info.

pub fn get_key_information(&self, key: PublicKey) -> Option<JsonKeyInfo>

Returns the JsonKeyInfo corresponding to a specific key

pub fn get_key_information_batch(
&self,
keys: Vec<PublicKey>
) -> Vec<Option<JsonKeyInfo>>

Returns the JsonKeyInfo corresponding to a specific key

pub fn get_drop_information(
&self,
drop_id: Option<u128>,
key: Option<PublicKey>
) -> JsonDrop

Returns the JsonDrop corresponding to a drop ID. If a key is specified, it will return the drop info for that key.

pub fn get_key_supply_for_drop(&self, drop_id: u128) -> u64

Returns the total supply of active keys for a given drop

pub fn get_keys_for_drop(
&self,
drop_id: u128,
from_index: Option<U128>,
limit: Option<u64>
) -> Vec<JsonKeyInfo>

Paginate through keys in a specific drop

pub fn get_drop_supply_for_owner(&self, account_id: AccountId) -> u64

Returns the total supply of active drops for a given owner

pub fn get_drops_for_owner(
&self,
account_id: AccountId,
from_index: Option<U128>,
limit: Option<u64>
) -> Vec<JsonDrop>

Return a vector of drop information for a owner

pub fn get_nft_supply_for_drop(&self, drop_id: u128) -> u64

Return the total supply of token IDs for a given drop

pub fn get_nft_token_ids_for_drop(
&self,
drop_id: u128,
from_index: Option<U128>,
limit: Option<u64>
) -> Vec<String>

Paginate through token IDs in a drop

pub fn get_next_drop_id(&self) -> u128

Returns the current nonce on the contract

pub fn get_fees_collected(&self) -> U128

Returns how many fees the contract has collected

pub fn get_gas_price(&self) -> U128

Returns the current GAS price stored on the contract

pub fn get_root_account(&self) -> String

Returns the current linkdrop contract

pub fn get_fees_per_user(&self, account_id: AccountId) -> Option<(U128, U128)>

Returns the current fees associated with an account

impl Keypom

pub fn ext(account_id: AccountId) -> KeypomExt

API for calling this contract’s functions in a subsequent execution.

impl Keypom

pub fn new(root_account: AccountId, owner_id: AccountId) -> Self

Initialize contract and pass in the desired deployed linkdrop contract (i.e testnet or near)