Struct keypom::KeypomExt

pub struct KeypomExt { /* private fields */ }

Implementations

impl KeypomExt

pub fn register_ft_contract(self, account_id: AccountId) -> Promise

Add a prohibited method to the list of methods that can’t be called by a FC Drop

pub fn set_contract_metadata(
self,
contract_metadata: ContractSourceMetadata
) -> Promise

Set the contract metadata with a spec and link

pub fn add_prohibited_method(self, method: String) -> Promise

Add a prohibited method to the list of methods that can’t be called by a FC Drop

pub fn set_root_account(self, root_account: AccountId) -> Promise

Set the desired linkdrop contract to interact with

pub fn set_fees(self, drop_fee: Option<U128>, key_fee: Option<U128>) -> Promise

Set the desired contract fees

pub fn set_fees_per_user(
self,
account_id: AccountId,
drop_fee: U128,
key_fee: U128
) -> Promise

Set the desired linkdrop contract to interact with

pub fn set_gas_price(self, yocto_per_gas: u128) -> Promise

Set the desired linkdrop contract to interact with

pub fn withdraw_fees(self, withdraw_to: AccountId) -> Promise

Withdraw the fees collected to the passed in Account Id

pub fn on_withdraw_fees(self, fees_collected: u128) -> Promise

Callback for withdrawing fees on the contract

impl KeypomExt

pub fn add_to_balance(self) -> Promise

pub fn withdraw_from_balance(self) -> Promise

pub fn get_user_balance(self, account_id: AccountId) -> Promise

impl KeypomExt

pub fn delete_keys(
self,
drop_id: u128,
public_keys: Option<Vec<PublicKey>>,
limit: Option<u8>,
delete_on_empty: Option<bool>
) -> Promise

pub fn refund_assets(
self,
drop_id: u128,
assets_to_refund: Option<u64>
) -> Promise

impl KeypomExt

pub fn create_drop(
self,
public_keys: Vec<PublicKey>,
passwords_per_use: Option<Vec<Option<Vec<JsonPasswordForUse>>>>,
passwords_per_key: Option<Vec<Option<String>>>,
deposit_per_use: U128,
config: Option<DropConfig>,
metadata: Option<String>,
ft_data: Option<JsonFTData>,
nft_data: Option<JsonNFTData>,
fc_data: Option<FCData>,
drop_id: Option<U128>
) -> Promise

pub fn add_keys(
self,
public_keys: Vec<PublicKey>,
passwords_per_use: Option<Vec<Option<Vec<JsonPasswordForUse>>>>,
passwords_per_key: Option<Vec<Option<String>>>,
drop_id: U128
) -> Promise

impl KeypomExt

pub fn ft_on_transfer(
self,
sender_id: AccountId,
amount: U128,
msg: U128
) -> Promise

Allows users to attach fungible tokens to the Linkdrops. Must have storage recorded by this point. You can only attach one set of FTs or NFT at a time.

pub fn ft_resolve_batch(
self,
amount: U128,
token_sender: AccountId,
token_contract: AccountId
) -> Promise

Self callback checks if fungible tokens were successfully transferred to the new account. If yes, do nothing. If no, refund original sender

pub fn ft_resolve_refund(self, drop_id: u128, num_to_refund: u64) -> Promise

Self callback checks if fungible tokens were successfully refunded. If yes, set keys registered to 0.

pub fn resolve_storage_check(
self,
public_keys: Vec<PublicKey>,
drop_id: u128,
required_deposit: u128
) -> Promise

self callback gets the storage balance bounds and inserts that into account data for each public key passed in

impl KeypomExt

pub fn nft_on_transfer(
self,
token_id: String,
sender_id: AccountId,
msg: U128
) -> Promise

pub fn nft_resolve_refund(
self,
drop_id: U128,
token_ids: Vec<String>
) -> Promise

self callback checks if NFT was successfully transferred to the new account. If yes, do nothing. If no, refund original sender

pub fn nft_resolve_transfer(
self,
token_id: String,
token_sender: AccountId,
token_contract: AccountId
) -> Promise

self callback checks if NFT was successfully transferred to the new account. If yes, do nothing. If no, refund original sender

impl KeypomExt

pub fn claim(self, account_id: AccountId, password: Option<String>) -> Promise

Claim tokens for specific account that are attached to the public key this tx is signed with.

pub fn create_account_and_claim(
self,
new_account_id: AccountId,
new_public_key: PublicKey,
password: Option<String>
) -> Promise

Create new account and and claim tokens to it.

pub fn on_claim_simple(
self,
owner_id: AccountId,
balance: U128,
storage_used: Balance,
auto_withdraw: bool
) -> Promise

self callback for simple linkdrops with no FTs, NFTs, or FCs.

pub fn on_claim_ft(
self,
account_id: AccountId,
owner_id: AccountId,
balance: U128,
storage_used: Balance,
ft_data: FTData,
execute: bool,
auto_withdraw: bool
) -> Promise

self callback for FT linkdrop

pub fn on_claim_nft(
self,
account_id: AccountId,
owner_id: AccountId,
balance: U128,
storage_used: Balance,
sender_id: AccountId,
contract_id: AccountId,
token_id: String,
execute: bool,
auto_withdraw: bool
) -> Promise

self callback for a linkdrop loaded with an NFT

pub fn on_claim_fc(
self,
account_id: AccountId,
owner_id: AccountId,
balance: U128,
storage_used: Balance,
fc_data: FCData,
drop_id: u128,
cur_key_id: u64,
remaining_uses: u64,
uses_per_key: u64,
execute: bool,
auto_withdraw: bool
) -> Promise

self callback checks if account was created successfully or not. If yes, refunds excess storage, sends NFTs, FTs etc..

impl KeypomExt

pub fn get_key_balance(self, key: PublicKey) -> Promise

Returns the balance associated with given key. This is used by the NEAR wallet to display the amount of the linkdrop

pub fn get_key_total_supply(self) -> Promise

Query for the total supply of keys on the contract

pub fn get_keys(self, from_index: Option<U128>, limit: Option<u64>) -> Promise

Paginate through all active keys on the contract and return a vector of key info.

pub fn get_key_information(self, key: PublicKey) -> Promise

Returns the JsonKeyInfo corresponding to a specific key

pub fn get_key_information_batch(self, keys: Vec<PublicKey>) -> Promise

Returns the JsonKeyInfo corresponding to a specific key

pub fn get_drop_information(
self,
drop_id: Option<u128>,
key: Option<PublicKey>
) -> Promise

Returns the JsonDrop corresponding to a drop ID. If a key is specified, it will return the drop info for that key.

pub fn get_key_supply_for_drop(self, drop_id: u128) -> Promise

Returns the total supply of active keys for a given drop

pub fn get_keys_for_drop(
self,
drop_id: u128,
from_index: Option<U128>,
limit: Option<u64>
) -> Promise

Paginate through keys in a specific drop

pub fn get_drop_supply_for_owner(self, account_id: AccountId) -> Promise

Returns the total supply of active drops for a given owner

pub fn get_drops_for_owner(
self,
account_id: AccountId,
from_index: Option<U128>,
limit: Option<u64>
) -> Promise

Return a vector of drop information for a owner

pub fn get_nft_supply_for_drop(self, drop_id: u128) -> Promise

Return the total supply of token IDs for a given drop

pub fn get_nft_token_ids_for_drop(
self,
drop_id: u128,
from_index: Option<U128>,
limit: Option<u64>
) -> Promise

Paginate through token IDs in a drop

pub fn get_next_drop_id(self) -> Promise

Returns the current nonce on the contract

pub fn get_fees_collected(self) -> Promise

Returns how many fees the contract has collected

pub fn get_gas_price(self) -> Promise

Returns the current GAS price stored on the contract

pub fn get_root_account(self) -> Promise

Returns the current linkdrop contract

pub fn get_fees_per_user(self, account_id: AccountId) -> Promise

Returns the current fees associated with an account

pub fn contract_source_metadata(self) -> Promise

Returns the current contract source metadata

impl KeypomExt

pub fn with_attached_deposit(self, amount: Balance) -> Self

pub fn with_static_gas(self, static_gas: Gas) -> Self

pub fn with_unused_gas_weight(self, gas_weight: u64) -> Self

impl KeypomExt

pub fn new(
self,
root_account: AccountId,
owner_id: AccountId,
contract_metadata: ContractSourceMetadata
) -> Promise

Initialize contract and pass in the desired deployed linkdrop contract (i.e testnet or near)